code/nba-player-loader/requirements.txt
requests==2.26.0
mysql-connector-python==8.0.28
